Use cart item data to provide initial values to min quantity and step values in the Cart API
Note that the usage of 'quantity_min' and 'quantity_step' is a bit cryptic here -- at this point there no method in WC_Cart or elsewhere that formalizes the use of these fields. The primary motivation to include them in the API comes from the need to validate them in a consistent manner when updating quantities (not handled yet).