Validate QuantitySelector input values on blur
This is an attempt to correctly handle qty input validation on blur. Currently, because validation was handled while typing in qty inputs, it was impossible for the user to see the typed value.
The need to decouple the value rendered by the Component from the initial qty value meant that we needed to introduce an additional property, resetQuantity, which onBlur falls back to when supplying an invalid value.