Fix Search Product placeholder's contrast
The placeholder's color of the Search Product block wasn't set in the
preview mode. The browser's user agent style was injected instead.

The placeholder's color in the latest version of Safari browser didn't
have enough contrast.

The solution was to set the right placeholder color instead of relying
on the browser default's color.