Revert icon control for inserter icon.
It was noted at the meetup that, by changing the inserter icon, we run
the risk of making it more confusing for the user when looking for the
block in the editor. They are likely accustomed to seeing the icon as it
is and would look for that familiar item  when trying to add the block.