Add text styling for simple products only
Due to the rich text capabilities of the Button block, the markup can look quite convoluted. Here are some examples of the inner HTML of `a` tag:
```html
- <strong><em><mark><code>Add to cart</code></mark></em></strong>
- <strong>Add</strong> <strong><em>to</em></strong> <em>cart</em>
```

This way merchant can style the text as they wish for the text shown for simple products.

For other product types, $product->add_to_cart_text() is a method that retrieves the text that should be displayed on the "Add to Cart" button.
For example, If the product is a variable product, text will be "Select options". If the product is out of stock, the text will be "Out of stock".

If the product is not a simple product, than we don't show any text styling because text styling is not supported for non-simple products because there is no way to style the text for non-simple products.