Fix Attribute terms sometimes being assigned children
Attribute terms (e.g. “Blue”, “Yellow”) have their own `id` property which
sometimes can overlap with their Taxonomy `id` (e.g. in the test case,
“Blue” had the same `id` as the “Size” taxonomy), and the function
building the tree representation for the `SearchListControl` would
erroneously assign children to it.

To fix this, we do an ugly patch here: we rename the `id` key of
`AttributeTerm` when changed to a `SearchListItem` to `termId` instead,
so we can differentiate. We do this because other items passed to
`SearchListControl` might possibly have several layers of children, but
we are sure that in the case of terms, we only have on level deep.

This may need a better refactoring at some point.